move pointer event type to input.d.ts
