+++ types/ext/settings.d.ts
@@ -15,7 +15,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import type {PointerType} from './text-scanner';
 import type * as Input from './input';
 
 export type OptionsContextFlag = 'clipboard';
@@ -28,7 +27,7 @@ export type OptionsContext1 = {
     flags?: OptionsContextFlag[];
     modifiers?: Input.Modifier[];
     modifierKeys?: Input.ModifierKey[];
-    pointerType?: PointerType;
+    pointerType?: Input.PointerType;
 };
 
 export type OptionsContext2 = {
@@ -39,7 +38,7 @@ export type OptionsContext2 = {
     flags?: OptionsContextFlag[];
     modifiers?: Input.Modifier[];
     modifierKeys?: Input.ModifierKey[];
-    pointerType?: PointerType;
+    pointerType?: Input.PointerType;
 };
 
 export type OptionsContext3 = {
@@ -50,7 +49,7 @@ export type OptionsContext3 = {
     flags?: OptionsContextFlag[];
     modifiers?: Input.Modifier[];
     modifierKeys?: Input.ModifierKey[];
-    pointerType?: PointerType;
+    pointerType?: Input.PointerType;
 };
 
 export type OptionsContext = OptionsContext1 | OptionsContext2 | OptionsContext3;
