+++ types/ext/text-scanner.d.ts
@@ -82,7 +82,7 @@ export type SentenceParsingOptions = {
 export type InputConfig = {
     include: string[];
     exclude: string[];
-    types: Set<PointerType>;
+    types: Set<Input.PointerType>;
     searchTerms: boolean;
     searchKanji: boolean;
     scanOnTouchMove: boolean;
@@ -100,8 +100,8 @@ export type InputConfig = {
 
 export type InputInfo = {
     input: InputConfig | null;
-    pointerType: PointerType;
-    eventType: PointerEventType;
+    pointerType: Input.PointerType;
+    eventType: Input.PointerEventType;
     passive: boolean;
     modifiers: Input.Modifier[];
     modifierKeys: Input.ModifierKey[];
@@ -187,35 +187,6 @@ export type Sentence = {
     text: string;
     offset: number;
 };
-
-export type PointerType = (
-    'pen' |
-    'mouse' |
-    'touch' |
-    'script'
-);
-
-export type PointerEventType = (
-    'mouseMove' |
-    'pointerOver' |
-    'pointerDown' |
-    'pointerMove' |
-    'pointerUp' |
-    'touchStart' |
-    'touchEnd' |
-    'touchMove' |
-    'click' |
-    'script'
-);
-
-/**
- * An enum representing the pen pointer state.
- * - `0` - Not active.
- * - `1` - Hovering.
- * - `2` - Touching.
- * - `3` - Hovering after touching.
- */
-export type PenPointerState = 0 | 1 | 2 | 3;
 
 export type SentenceTerminatorMap = Map<string, [includeCharacterAtStart: boolean, includeCharacterAtEnd: boolean]>;
